Move MissingEnvVarsError function to index file
